+++ services/__init__.py
@@ -1,6 +1,10 @@
 """Service layer initialization and exports."""
 
-from services.interfaces import MacroServiceInterface, SignalServiceInterface
+from services.interfaces import (
+    MacroServiceInterface,
+    RegimeServiceInterface,
+    SignalServiceInterface,
+)
 from services.macro_regime_service import MacroRegimeService
 from services.macro_service import MacroService
 from services.signal_service import SignalService
@@ -8,6 +12,7 @@ from services.signal_service import SignalService
 __all__ = [
     "MacroServiceInterface",
     "SignalServiceInterface",
+    "RegimeServiceInterface",
     "MacroService",
     "MacroRegimeService",
     "SignalService",